switched back return types to optimal size

+++ src/problems/smallest_multiple.rs
@@ -1,8 +1,8 @@
 // https://projecteuler.net/problem=5
 pub const PROBLEM_ID: u16 = 5;
 
-pub fn run() -> u64 {
-    let mut multiple: u64 = 20;
+pub fn run() -> u32 {
+    let mut multiple: u32 = 20;
 
     loop {
         if (1..=20).all(|x| -> bool { return multiple % x == 0 }) {
